Cleanup: use int for view_id's & view count
`size_t` is useful for memory sizes or offsets,
the number of views wont realistically exceed an int.

+++ source/blender/blenkernel/BKE_scene.h
@@ -153,17 +153,17 @@ bool        BKE_scene_multiview_is_stereo3d(const struct RenderData *rd);
 bool        BKE_scene_multiview_is_render_view_active(const struct RenderData *rd, const struct SceneRenderView *srv);
 bool        BKE_scene_multiview_is_render_view_first(const struct RenderData *rd, const char *viewname);
 bool        BKE_scene_multiview_is_render_view_last(const struct RenderData *rd, const char *viewname);
-size_t      BKE_scene_multiview_num_views_get(const struct RenderData *rd);
+int         BKE_scene_multiview_num_views_get(const struct RenderData *rd);
 struct SceneRenderView *BKE_scene_multiview_render_view_findindex(const struct RenderData *rd, const int view_id);
 const char *BKE_scene_multiview_render_view_name_get(const struct RenderData *rd, const int view_id);
-size_t      BKE_scene_multiview_view_id_get(const struct RenderData *rd, const char *viewname);
+int         BKE_scene_multiview_view_id_get(const struct RenderData *rd, const char *viewname);
 void        BKE_scene_multiview_filepath_get(struct SceneRenderView *srv, const char *filepath, char *r_filepath);
 void        BKE_scene_multiview_view_filepath_get(const struct RenderData *rd, const char *filepath, const char *view, char *r_filepath);
 const char *BKE_scene_multiview_view_suffix_get(const struct RenderData *rd, const char *viewname);
-const char *BKE_scene_multiview_view_id_suffix_get(const struct RenderData *rd, const size_t view_id);
+const char *BKE_scene_multiview_view_id_suffix_get(const struct RenderData *rd, const int view_id);
 void        BKE_scene_multiview_view_prefix_get(struct Scene *scene, const char *name, char *rprefix, const char **rext);
 void        BKE_scene_multiview_videos_dimensions_get(const struct RenderData *rd, const size_t width, const size_t height, size_t *r_width, size_t *r_height);
-size_t      BKE_scene_multiview_num_videos_get(const struct RenderData *rd);
+int         BKE_scene_multiview_num_videos_get(const struct RenderData *rd);
 
 #ifdef __cplusplus
 }
